+++ Hazel/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -73,7 +73,7 @@ void ImGuiLayer::End()
 {
 	ImGuiIO& io = ImGui::GetIO();
 	Application& app = Application::Get();
-	io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+	io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 	// Rendering
 	ImGui::Render();
@@ -92,5 +92,8 @@ void ImGuiLayer::OnImGuiRender()
 {
 	static bool show = true;
 	ImGui::ShowDemoWindow(&show);
+
+	//ImGui::Begin("");
+	//ImGui::End();
 }
 }
